PagedScrollView stub.
