+++ Corund/Corund.csproj
@@ -141,8 +141,10 @@
     <Compile Include="Visuals\TextObject.cs" />
     <Compile Include="Visuals\UI\Button.cs" />
     <Compile Include="Visuals\UI\CheckButton.cs" />
+    <Compile Include="Visuals\UI\PagedScrollView.cs" />
     <Compile Include="Visuals\UI\ScrollDirection.cs" />
     <Compile Include="Visuals\UI\ScrollView.cs" />
+    <Compile Include="Visuals\UI\ScrollViewBase.cs" />
     <Compile Include="Visuals\UI\StaticGridPanel.cs" />
     <Compile Include="Visuals\UI\StackPanel.cs" />
   </ItemGroup>
